Move direct NMS code into the correct package
